+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.6.3
 - Fixed plugin version being 1.5.0 for some reason...
+- Updated EvolveRandomly to now modify multiple sets of cards for random evolution.
 
 ## v1.6.2
 - Created `Randomly Evolve` - A card bearing this sigil will grow into a random form after 1 turn on the board.